Re-organise aiohttp code
The main thing I'm doing here is moving all of the asyncio code
out of the Client classes so that it is completely encapsulated by
`proxy_single_request()` and `proxy_multiple_requests()`.
This is helpful because it means for testing we don't need to deal
with async mocking. We can just fake a return value for
`proxy_single_request()` or `proxy_multiple_requests()`.

The `Request` object just allows us to easily encapsulate
a URL and some `GET` params together, which is a nice way to
keep them together if we've got an array of them.