Use existing metadata key for cancellation reason
This reverts a recent change to add a key `cancellation_reason` to the
ballot.

The initial change was intended to expose the detail of the
cancellation, as calculated form the number of seats and candidates.

When Sym went to implement this change on a project, he realised that
this was actually more confusing, as in some cases the consumer had to
look at the `metadata` object and in others they had to look at the
`cancellation_reason` key.

The change in this commit moves all reasons into a `metadata` object.

If the election is cancelled and we don't have existing metadata for it,
we create a `metadata` dict and add it to the ballot.