Be more clever when showing boundary changes
Context - gss codes have changed in a lot of places where the boundaries
of new constituencies haven't changed. In addition boundaries have
changed in ways that don't effect who votes - i.e. rivers, high tide
lines, chunks of moorland etc.

Solution - Take the new constituency which has greatest overlap to each
old constituency. Take the symmetrical difference of these, and see how
many properties fall in that differnce. If the number of properties is
below some threshold (50), then declare `NO_CHANGE` for lookups with
that old/new pair. This means that the edge cases - i.e. the small
number of addresses that are in only one out of the old/new pair will be
told about the big change that has effected them.